fatafat 218 test findings
1) The StartMetadataAPI was intercepting the flow in the wrong place in order to try out the alternate parsing
for the jpmml add command
2) The models were moved into subdirectories according to their origin (KNIME and Rattle)
3) The Iris dataset was incorrect... version present now has partition key value
4) The addition of a second application to the MetadataAPI (the test harness for the AlternateParser) confused
the jvm when trying to run command.  A "mainClass in assembly := Some("com.ligadata.MetadataAPI.StartMetadataAPI")
was added to the build.sbt to prefer it.
5) Comment corrected in object MdMgr for one of the format routines.

Status:
The commands are being parsed and the MetadataAPIImpl.AddModel and AddJpmmlModel are being invoked, but there is
a problem with the format of the version for the model I submitted.  I need to format it before giving to
the function.

Once by this issue, I believe the metadata for the Jpmml model will be cataloged. After that, it is testing the use
of the ModelDef and the JpmmlAdapter factory that manages the instance cache for the models.